data_compressor/requirements.txt
Pillow
opencv-python
optipng
